% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{duckr_to_parquet}
\alias{duckr_to_parquet}
\title{Export a table or view to a Parquet file}
\usage{
duckr_to_parquet(name, file, dir = ".", con = duckr_con(), overwrite = FALSE)
}
\arguments{
\item{name}{Name of the table or view to export.}

\item{file}{Output Parquet file name.}

\item{dir}{Directory to write \code{file} into. Defaults to \code{"."}.}

\item{con}{A DBI connection. Defaults to the current connection.}

\item{overwrite}{If \code{FALSE} (default) error when the output file exists; if
\code{TRUE} overwrite it.}
}
\value{
The connection \code{con}, invisibly.
}
\description{
Copies the content of a table or view of the connection to a Parquet file.
}
\examples{
con <- duckr_connect()
duckr_add_df(mtcars, name = "cars")
duckr_to_parquet("cars", "cars.parquet", dir = tempdir())
duckr_close()
}
\seealso{
Other data exporters:
\code{\link[=duckr_to_csv]{duckr_to_csv()}}
}
\concept{data exporters}
