% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{duckr_add_parquet}
\alias{duckr_add_parquet}
\title{Add a Parquet file as a view or table}
\usage{
duckr_add_parquet(
  file,
  dir = ".",
  name,
  con = duckr_con(),
  materialize = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{file}{Parquet file name.}

\item{dir}{Directory containing \code{file}. Defaults to \code{"."}.}

\item{name}{Name of the object to create.}

\item{con}{A DBI connection. Defaults to the current connection.}

\item{materialize}{If \code{FALSE} (default) create a \code{VIEW}; if \code{TRUE} create a
\code{TABLE}.}

\item{overwrite}{If \code{FALSE} (default) error when the object exists; if \code{TRUE}
use \verb{CREATE OR REPLACE}.}
}
\value{
The connection \code{con}, invisibly.
}
\description{
Add a Parquet file as a view or table
}
